Fronts.ParamEstim module: parameter estimation support

The ParamEstim submodule provides support for optimization-based parameter estimation runs using Fronts.

Fronts.ParamEstim.RSSCostFunction — Type

RSSCostFunction{fit_D0}(func, ϕ, data[, weights; catch_errors, D0tol, ϕi_hint])

Residual sum of squares cost function for parameter estimation.

Type parameters

fit_D0::Bool: whether to fit an additional constant factor D0 that affects the diffusivity. Values

of D0 can be found with relative efficiency without additional solver calls; so if any such constant factors affecting the diffusivity are unknown, it is recommended not to fit those factors directly but set fit_D0 to true instead. Values of D0 are found internally by local optimization, and they can be retrieved by calling the candidate function.

Arguments

func: function that takes a vector of parameter values and returns either a Fronts.Solution or a

Fronts.Problem. If func returns a Problem, it is solved with trysolve. func is also allowed to return nothing to signal that no solution could be found for the parameter values, which will imply an infinite cost (see also the catch_errors keyword argument).

ϕ: vector of values of the Boltzmann variable. See Fronts.ϕ.

data: data to fit. Must be a vector of the same length as ϕ.

weights: optional weights for the data. If given, must be a vector of the same length as data.

Keyword arguments

catch_errors=(Fronts.SolvingError,): collection of exception types that func is allowed to throw;

any of these exceptions will be caught and will result in an infinite cost.

D0tol=1e-3: if fit_D0 is true, a tolerance for D0.

ϕi_hint=ϕ[end]: if fit_D0 is true, a hint as to the point in ϕ where the initial condition begins.

The hint will be used as an aid in finding the optimal value for D0.

(::RSSCostFunction)(p::AbstractVector)

Return the cost of the solution obtained with parameter values p.

The RSSCostFunction object is meant to be passed to your optimizer of choice for minimization as the objective function.

If you need to know more than just the cost, call the candidate function instead.

See also: candidate, Fronts.Solution, Fronts.Problem, trysolve

Fronts.ParamEstim.candidate — Function

candidate(cf::RSSCostFunction, ::AbstractVector)
candidate(cf::RSSCostFunction, ::Fronts.Problem)
candidate(cf::RSSCostFunction, ::Fronts.Solution)
candidate(cf::RSSCostFunction, ::Nothing)

Return the candidate solution (including the cost) for a given cost function and parameter values, problem, or solution.

The return of this function has the following fields:

sol: the solution, or nothing if no solution could be found.

D0: if cf has fit_D0 set to true and sol is not nothing, the found value of D0.

cost: the cost of the solution; infinite if sol is nothing.

Fronts.ParamEstim.trysolve — Function

trysolve(prob[, catch_errors, kwargs...])::Union{Fronts.Solution, Nothing}

Attempt to solve a problem with Fronts.solve and return the solution, but catch any exceptions of the types included in catch_errors and return nothing on such failures.

Arguments

prob: problem to be solved.

Keyword arguments

catch_errors=(Fronts.SolvingError,): collection of exception types that should be caught.

kwargs...: any additional keyword arguments are passed to solve.

See also: Fronts.solve, Fronts.SolvingError