The Boltzmann transformation

Lower-level API to work with the Boltzmann transformation.

Fronts.ϕ — Function

Fronts.ϕ(r, t)

Evaluate the Boltzmann variable ϕ at position r and time t.

Type \phi<tab> to obtain the ϕ symbol.

The Boltzmann variable is defined as ϕ=r/√t and makes the Boltzmann transformation possible.

To prevent possible name clashes, this function is not exported.

See also: transform

Fronts.∂ϕ_∂r — Function

∂ϕ_∂r(r, t)

Partial derivative of the Boltzmann variable.

Type \partial<tab> to obtain the ∂ symbol; \phi<tab> to obtain the ϕ symbol.

See also: ϕ

Fronts.∂ϕ_∂t — Function

∂ϕ_∂t(r, t)

Partial derivative of the Boltzmann variable.

Type \partial<tab> to obtain the ∂ symbol; \phi<tab> to obtain the ϕ symbol.

See also: ϕ

Fronts.r — Function

Fronts.r(ϕ, t)

Convert back from the Boltzmann variable to r.

To prevent possible name clashes, this function is not exported.

See also: ϕ

Fronts.t — Function

Fronts.t(ϕ, r)

Convert back from the Boltzmann variable to t.

To prevent possible name clashes, this function is not exported.

See also: ϕ

Fronts.transform — Function

transform(r, t)

Same as ϕ(r,t).

See also: ϕ

transform(eq::Equation) -> DifferentialEquations.ODEFunction

Transform eq into an ordinary differential equation (ODE) defined in terms of the Boltzmann variable ϕ.

Returns an ODE with independent variable ϕ and two components, where the first is the solution itself and the second component is the ϕ-derivative of the solution. The ODE is optimized for components stored in StaticArrays.SVectors.

See also: DifferentialEquations, StaticArrays.SVector

transform(prob::CauchyProblem) -> DifferentialEquations.ODEProblem

Transform prob into an ODE problem in terms of ϕ. The ODE problem is set up to terminate automatically (.retcode == :Terminated) when the steady state is reached.

See also: DifferentialEquations

Fronts.TransformedFunction — Type

TransformedFunction

Abstract type for functions of the Boltzmann variable ϕ.

Every subtype of TransformedFunction gets access to the following methods:

(::TransformedFunction)(r, t)
d_dϕ(::TransformedFunction, r, t)
∂_∂r(::TransformedFunction, r, t)
∂_∂t(::TransformedFunction, r, t)

Implementation

In order to access the previous methods, a type T <: TransformedFunction must define these methods:

(::T)(ϕ)
d_dϕ(::T, ϕ)