Equations

Fronts.DiffusionEquation — Type

DiffusionEquation(D; symbol=:θ)
DiffusionEquation{m}(D; symbol=:θ)

Nonlinear diffusion equation.

Arguments

D: diffusivity function.

Keyword arguments

symbol::Symbol=:θ: optional symbol used to represent the unknown function in the output.

Type parameters

m::Int=1: number of spatial dimensions:

1 for non-radial one-dimensional diffusion (default);

2 for radial diffusion in polar or cylindrical coordinates;

3 for radial diffusion in spherical coordinates.

Examples

julia> D(θ) = θ^4
D (generic function with 1 method)

julia> eq = Fronts.DiffusionEquation(D)
∂θ/∂t = ∂(D(θ)*∂θ/∂r)/∂r

julia> eq = Fronts.DiffusionEquation{2}(D)
∂θ/∂t = 1/r*∂(r*D(θ)*∂θ/∂r)/∂r

julia> eq = Fronts.DiffusionEquation{3}(D, symbol=:c)
∂c/∂t = 1/r²*∂(r²*D(c)*∂c/∂r)/∂r

Fronts.isindomain — Function

isindomain(eq::DiffusionEquation, θ) -> Bool

true if eq is well defined for the solution value θ; false otherwise.