Problems in finite domains

Fronts.FiniteProblem — Type

FiniteProblem{Eq<:Equation}

Abstract type for problems defined in finite domains.

Fronts.FiniteDirichletProblem — Type

FiniteDirichletProblem(eq, rstop; i, b) <: FiniteProblem

Models eq in the domain 0 ≤ r ≤ rstop with initial condition i and boundary condition b.

Fronts.FiniteFluxProblem — Type

FiniteFluxProblem(eq, rstop; i, qb) <: FiniteProblem

Models eq in the domain 0 ≤ r ≤ rstop with initial condition i and flux boundary condition qb.

Fronts.FiniteReservoirProblem — Type

FiniteReservoirProblem(eq, rstop; i, b, capacity) <: FiniteProblem

Models eq in the domain 0 ≤ r ≤ rstop with initial condition i and a reservoir boundary condition with capacity capacity.

Fronts.solve — Method

solve(prob::FiniteProblem{<:DiffusionEquation{1}}, tstop[; N, tol, Δt]) -> FiniteSolution

Solve the finite problem prob up to time tstop.

Uses a finite difference scheme with backward Euler time discretization and a second-order central difference scheme for the fluxes.

Keyword arguments

N=500: number of points in the spatial grid.

tol=1e-3: absolute tolerance for the solution.

Δt=1: initial time step.

Fronts.FiniteSolution — Type

Solution to a finite problem.

(::FiniteSolution)(r, t)

Evaluate the solution at location r and time t.