Evaluating solutions

Fronts.Solution — Type

Solution to a problem.

(::Solution)(r, t)
(::Solution)(o)

Evaluate the solution.

Properties

retcode: termination status of the solver. See ReturnCode.

i: initial value.

b: boundary value.

d_dob: o-derivative at the boundary, where o is the Boltzmann variable. See also o.

ob: boundary constant. See also rb.

oi: for o≥oi, the solution evaluates to the initial value.

prob: problem solved.

alg: algorithm used.

original: original solution object, if applicable.

Fronts.d_dr — Function

d_dr(::Solution, r, t)

Spatial derivative of the solution.

d_dr(::Solution, :b, t)

Spatial derivative of the solution at the boundary.

Fronts.d_dt — Function

d_dt(::Solution, r, t)

Time derivative of the solutio.

d_dt(::Solution, :b, t)

Time derivative of the solution at the boundary.

Fronts.flux — Function

flux(::Solution, r, t)

Flux.

flux(::Solution, :b, t)

Boundary flux.

Fronts.d_do — Function

d_do(::Solution, r, t)
d_do(::Solution, o)

o-derivative of the solution, where o is the Boltzmann variable.

See also: o

Fronts.rb — Function

rb(::Solution, t)

Location of the boundary in the solution at time t, equal to ob*√t.

Fronts.sorptivity — Method

sorptivity(::Solution)

Sorptivity.

sorptivity(::Solution, o)

Sorptivity, computed from the given value of o.

References

PHILIP, J. R. The theory of infiltration: 4. Sorptivity and algebraic infiltration equations. Soil Science, 1957, vol. 84, no. 3, p. 257-264.

Fronts.sorptivity — Method

sorptivity(::Solution)

Sorptivity.

sorptivity(::Solution, o)

Sorptivity, computed from the given value of o.

References

PHILIP, J. R. The theory of infiltration: 4. Sorptivity and algebraic infiltration equations. Soil Science, 1957, vol. 84, no. 3, p. 257-264.

Fronts.FiniteSolution — Type

Solution to a finite problem.

(::FiniteSolution)(r, t)

Evaluate the solution at location r and time t.