Solving problems

Fronts.solve — Function

solve(prob::DirichletProblem[; itol, maxiter, d_dϕb_hint]) -> Solution
solve(prob::FlowrateProblem[; itol, ϕbtol, maxiter, b_hint]) -> Solution
solve(prob::CauchyProblem) -> Solution

Solve the problem prob.

Keyword arguments

itol=1e-3: absolute tolerance for the initial condition.

ϕbtol=1e-6: maximum tolerance for ϕb. Allows solving FlowrateProblems with boundaries at r=0.

maxiter=100: maximum number of iterations.

d_dϕb_hint, b_hint: optional hints for the algorithms.

Type \phi<tab> to obtain the ϕ symbol.

Exceptions

This function throws an SolvingError if an acceptable solution is not found (within the maximum number of iterations, if applicable). However, in situations where solve can determine that the problem is "unsolvable" before the attempt to solve it, it will signal this by throwing a DomainError instead. Other invalid argument values will raise ArgumentErrors.

See also: Solution, SolvingError

Fronts.solve — Method

solve(::DirichletProblem{<:DiffusionEquation{1}}, ::MathiasAndSander[; maxiter]) -> Solution

Solve a Dirichlet problem using the pseudospectral method of Mathias and Sander (2021).

Keyword arguments

maxiter: maximum number of iterations.

References

MATHIAS, S. A.; SANDER, G. C. Pseudospectral methods provide fast and accurate solutions for the horizontal infiltration equation. Journal of Hydrology, 2021, vol. 598, p. 126407.

See also: MathiasAndSander, Solution, SolvingError

Fronts.SolvingError — Type

Exception thrown when solve fails to find a solution.

Fronts.MathiasAndSander — Type

MathiasAndSander([; N, Ftol])

Pseudospectral method of Mathias and Sander (2021).

Keyword arguments

N=100: number of Chebyshev nodes.

Ftol=1e-6: tolerance for the flux–concentration relationship.

References

MATHIAS, S. A.; SANDER, G. C. Pseudospectral methods provide fast and accurate solutions for the horizontal infiltration equation. Journal of Hydrology, 2021, vol. 598, p. 126407.

See also: solve