The Boltzmann transformation

Lower-level API to work with the Boltzmann transformation.

Fronts.o — Function

Fronts.o(r, t)

Evaluate the Boltzmann variable o at position r and time t.

The Boltzmann variable is defined as o=r/√t and makes the Boltzmann transformation possible.

To prevent possible name clashes, this function is not exported.

See also: boltzmann

Fronts.do_dr — Function

do_dr(r, t)

Partial derivative of the Boltzmann variable.

See also: o

Fronts.do_dt — Function

do_dt(r, t)

Partial derivative of the Boltzmann variable.

See also: o

Fronts.r — Function

Fronts.r(o, t)

Convert back from the Boltzmann variable to r.

To prevent possible name clashes, this function is not exported.

See also: o

Fronts.t — Function

Fronts.t(o, r)

Convert back from the Boltzmann variable to t.

To prevent possible name clashes, this function is not exported.

See also: o

Fronts.boltzmann — Function

boltzmann(r, t)

Same as o(r,t).

See also: o

boltzmann(eq::DiffusionEquation) -> DifferentialEquations.ODEFunction

Transform eq into an ordinary differential equation (ODE) defined in terms of the Boltzmann variable o.

Returns an ODE with independent variable o and two components, where the first is the solution itself and the second component is the o-derivative of the solution. The ODE is optimized for components stored in StaticArrays.SVectors.

See also: DifferentialEquations, StaticArrays.SVector

boltzmann(prob::CauchyProblem) -> DifferentialEquations.ODEProblem
boltzmann(prob::SorptivityCauchyProblem) -> DifferentialEquations.ODEProblem

Transform prob into an ODE problem in terms of the Boltzmann variable o.

The ODE problem is set up to terminate automatically (with .retcode == ReturnCode.Success) when the steady state is reached.

See also: DifferentialEquations