Equations

Fronts.Equation — Type

abstract type Equation{m} end

Abstract supertype for equations that can be solved with this package.

Type parameters

m: number of spatial dimensions:

1 for a non-radial one-dimensional equation (default);

2 for a radial equation in polar or cylindrical coordinates;

3 for a radial equation in spherical coordinates.

Fronts.DiffusionEquation — Type

DiffusionEquation(D; symbol=:θ) <: Equation{1}
DiffusionEquation{m}(D; symbol=:θ) <: Equation{m}

Nonlinear diffusion equation.

DiffusionEquation(model) <: Equation{1}
DiffusionEquation{m}(model) <: Equation{m}

Nonlinear diffusion equation describing flow in a porous medium, with the diffusivity defined by a porous model.

Arguments

D: diffusivity function.

model::PorousModels.UnsaturatedFlowModel: unsaturated flow model from which to obtain the diffusivity function.

Keyword arguments

symbol::Symbol=:θ: optional symbol used to represent the unknown function in the output.

Type parameters

m::Int=1: number of spatial dimensions:

1 for non-radial one-dimensional diffusion (default);

2 for radial diffusion in polar or cylindrical coordinates;

3 for radial diffusion in spherical coordinates.

Examples

julia> D(θ) = θ^4
D (generic function with 1 method)

julia> eq = Fronts.DiffusionEquation(D)
∂θ/∂t = ∂(D(θ)*∂θ/∂r)/∂r

julia> eq = Fronts.DiffusionEquation{2}(D)
∂θ/∂t = 1/r*∂(r*D(θ)*∂θ/∂r)/∂r

julia> eq = Fronts.DiffusionEquation{3}(D, symbol=:c)
∂c/∂t = 1/r²*∂(r²*D(c)*∂c/∂r)/∂r

See also: PorousModels.UnsaturatedFlowModel

Fronts.RichardsEquation — Type

RichardsEquation(; C, K, symbol=:h) <: Equation{1}
RichardsEquation{m}(; C, K, symbol=:h) <: Equation{m}

Horizontal Richards equation, pressure-based formulation.

RichardsEquation(model) <: Equation{1}
RichardsEquation{m}(model) <: Equation{m}

Horizontal Richards equation, pressure-based formulation, with properties defined by a porous model.

Arguments

pm::PorousModels.UnsaturatedFlowModel: unsaturated flow model from which to obtain the relevant functions.

Keyword arguments

C: hydraulic capacity function, defined in terms of the unknown.

K: hydraulic conductivity function, defined in terms of the unknown.

symbol::Symbol=:h: optional symbol used to represent the unknown function in the output.

Type parameters

m::Int=1: number of spatial dimensions:

1 for non-radial one-dimensional flow (default);

2 for radial flow in polar or cylindrical coordinates;

3 for radial flow in spherical coordinates.

See also: PorousModels.UnsaturatedFlowModel

Fronts.isindomain — Function

isindomain(eq::Equation, val) -> Bool

true if eq is well defined for the solution value val; false otherwise.

Fronts.diffusivity — Function

diffusivity(eq::Equation, val)

Diffusivity of the solution variable of eq with value val.

Fronts.flow_diffusivity — Function

flow_diffusivity(eq::Equation, val)

Diffusivity of eq at val, as used to determine flow-related quantities (e.g. flux and sorptivity).

Implementation

Delegates to diffusivity by default.