Inverse problems

Fronts.InverseProblem — Type

InverseProblem(o, θ[, weights; i, b, ob])

Problem type for inverse functions and parameter estimation with experimental data.

Arguments

o::AbstractVector: values of the Boltzmann variable. See o.

θ::AbstractVector: observed solution values at each point in o.

weights: optional weights for the data.

Keyword arguments

i: initial value, if known.

b: boundary value, if known.

ob=0: value of o at the boundary.

See also

diffusivity, sorptivity, Fronts.ParamEstim

Fronts.inverse — Function

inverse(prob::InverseProblem) -> Function

inverse(o, θ) -> Function

Extract a diffusivity function D from a solution to a semi-infinite one-dimensional nonlinear diffusion problem, where the solution is given as a set of discrete points.

Interpolates the given solution with a PCHIP monotonic spline and uses the Bruce and Klute method to reconstruct D.

Due to the method used for interpolation, D will be continuous but will have discontinuous derivatives.

Arguments

prob::InverseProblem: inverse problem. See InverseProblem.

o::AbstractVector: values of the Boltzmann variable. See o.

θ::AbstractVector: solution values at each point in o.

References

GERLERO, G. S.; BERLI, C. L. A.; KLER, P. A. Open-source high-performance software packages for direct and inverse solving of horizontal capillary flow. Capillarity, 2023, vol. 6, no. 2, p. 31-40.

BRUCE, R. R.; KLUTE, A. The measurement of soil moisture diffusivity. Soil Science Society of America Journal, 1956, vol. 20, no. 4, p. 458-462.

Fronts.sorptivity — Method

sorptivity(::InverseProblem)
sorptivity(o, θ)

Calculate the sorptivity of a solution to a semi-infinite one-dimensional nonlinear diffusion problem, where the solution is given as a set of discrete points.

Uses numerical integration.

Arguments

prob::InverseProblem: inverse problem. See InverseProblem.

o::AbstractVector: values of the Boltzmann variable. See o.

θ::AbstractVector: solution values at each point in o.

Keyword arguments

i=nothing: initial value. If nothing, the initial value is taken from θ[end].

b=nothing: boundary value. If nothing, the boundary value is taken from θ[begin].

ob=0: value of o at the boundary.

References

PHILIP, J. R. The theory of infiltration: 4. Sorptivity and algebraic infiltration equations. Soil Science, 1957, vol. 83, no. 5, p. 345-357.