Problems

Fronts.DirichletProblem — Type

DirichletProblem(eq; i=<initial value>, b=<boundary value>, ϕb=0)
DirichletProblem(D; i=<initial value>, b=<boundary value>, ϕb=0)

Semi-infinite problem with a Dirichlet boundary condition.

Arguments

eq::Equation: governing equation.

D: diffusivity function. Shortcut for DirichletProblem(DiffusionEquation(D), ...).

Keyword arguments

i: initial value.

b: imposed boundary value.

ϕb=0 (\phi<tab>b): boundary constant for an optional moving boundary.

At time t, the boundary is located at ϕb*√t. Must be positive if eq is radial.

Examples

julia> D(θ) = θ^4
D (generic function with 1 method)

julia> prob = Fronts.DirichletProblem(D, i=1, b=2)
⎧ ∂θ/∂t = ∂(D(θ)*∂θ/∂r)/∂r, r>0,t>0
⎨ θ(r,0) = 1, r>0
⎩ θ(0,t) = 2, t>0

See also: Equation

Fronts.FlowrateProblem — Type

FlowrateProblem(eq; i=<initial value>, Qb=<boundary flowrate>, angle=2π, height=1, ϕb=0)

Semi-infinite radial (polar/cylindrical) problem with an imposed-flowrate boundary condition.

Arguments

eq::Equation{2}: governing equation.

Keyword arguments

i: initial value.

Qb: imposed boundary flowrate.

angle=2π: total angle covered by the domain.

height=1: domain height.

ϕb=0 (\phi<tab>b): boundary constant for an optional moving boundary.

At time t, the boundary is located at ϕb*√t.

Examples

julia> D(θ) = θ^4
D (generic function with 1 method)

julia> eq = Fronts.DiffusionEquation{2}(D)
∂θ/∂t = 1/r*∂(r*D(θ)*∂θ/∂r)/∂r

julia> prob = Fronts.FlowrateProblem(eq, i=1, Qb=1)
⎧ ∂θ/∂t = 1/r*∂(r*D(θ)*∂θ/∂r)/∂r, r>0,t>0
⎨ θ(r,0) = 1, r>0
⎩ Qb(0,t) = 1, t>0

See also: Equation

Fronts.CauchyProblem — Type

CauchyProblem(eq; b=<boundary value>, d_dϕb=<boundary ϕ-derivative>, ϕb=0)
CauchyProblem(D; b=<boundary value>, d_dϕb=<boundary ϕ-derivative>, ϕb=0)

Semi-infinite problem with a Cauchy boundary condition (and unknown initial condition).

Arguments

eq::Equation: governing equation.

D: diffusivity function. Shortcut for CauchyProblem(DiffusionEquation(D), ...).

Keyword arguments

b: imposed boundary value.

d_dϕb: imposed value of the ϕ-derivative of the solution at the boundary, where ϕ is the Boltzmann variable.

This value is equivalent to √t*∂_∂r(<solution>, :b, t) at any time t>0.

ϕb=0 (\phi<tab>b): boundary constant for an optional moving boundary.

At time t, the boundary is located at ϕb*√t. Must be positive if eq is radial.

Examples

julia> D(θ) = θ^4
D (generic function with 1 method)

julia> prob = Fronts.CauchyProblem(D, b=2, d_dϕb=-0.1)
⎧ ∂θ/∂t = ∂(D(θ)*∂θ/∂r)/∂r, r>0,t>0
⎨ θ(0,t) = 2, t>0
⎩ √t*∂θ/∂r(0,t) = -0.1, t>0

See also: Equation

Fronts.monotonicity — Function

monotonicity(prob) -> Int

Whether the solution to prob must be decreasing (-1), constant (0) or increasing (+1) in r.