Problems

Fronts.Problem — Type

abstract type Problem{Eq<:Equation} end

Abstract supertype for problems that can be solved with this package.

Type parameters

Eq: type of the governing equation

See also: Equation

Fronts.DirichletProblem — Type

DirichletProblem(eq; i, b[, ob]) <: Problem{typeof(eq)}
DirichletProblem(D; i, b[, ob]) <: Problem{DiffusionEquation{1}}

Semi-infinite problem with a Dirichlet boundary condition.

Arguments

eq::Equation: governing equation.

D: diffusivity function. Shortcut for DirichletProblem(DiffusionEquation(D), ...).

Keyword arguments

i: initial value.

b: imposed boundary value.

ob=0: boundary constant for an optional moving boundary. At time t, the boundary is located at ob*√t. Must be positive if eq is radial.

Examples

julia> D(θ) = θ^4
D (generic function with 1 method)

julia> prob = Fronts.DirichletProblem(D, i=1, b=2)
⎧ ∂θ/∂t = ∂(D(θ)*∂θ/∂r)/∂r, r>0,t>0
⎨ θ(r,0) = 1, r>0
⎩ θ(0,t) = 2, t>0

See also: Equation

Fronts.FlowrateProblem — Type

FlowrateProblem(eq; i, Qb[, angle, height, ob]) <: Problem{typeof(eq)}

Semi-infinite radial (polar/cylindrical) problem with an imposed-flowrate boundary condition.

Arguments

eq::Equation{2}: governing equation.

Keyword arguments

i: initial value.

Qb: imposed boundary flowrate.

angle=2π: total angle covered by the domain.

height=1: domain height.

ob=0: boundary constant for an optional moving boundary. At time t, the boundary is located at ob*√t.

Examples

julia> D(θ) = θ^4
D (generic function with 1 method)

julia> eq = Fronts.DiffusionEquation{2}(D)
∂θ/∂t = 1/r*∂(r*D(θ)*∂θ/∂r)/∂r

julia> prob = Fronts.FlowrateProblem(eq, i=1, Qb=1)
⎧ ∂θ/∂t = 1/r*∂(r*D(θ)*∂θ/∂r)/∂r, r>0,t>0
⎨ θ(r,0) = 1, r>0
⎩ Qb(0,t) = 1, t>0

See also: Equation

Fronts.CauchyProblem — Type

CauchyProblem(eq; b, d_dob[, ob]) <: Problem{typeof(eq)}
CauchyProblem(D; b, d_dob[, ob]) <: Problem{DiffusionEquation{1}}

Semi-infinite problem with a Cauchy boundary condition (and unknown initial condition).

Arguments

eq::Equation: governing equation.

D: diffusivity function. Shortcut for CauchyProblem(DiffusionEquation(D), ...).

Keyword arguments

b: imposed boundary value.

d_dob: imposed value of the o-derivative of the solution at the boundary, where o is the Boltzmann variable.

This value is equivalent to √t*d_dr(<solution>, :b, t) at any time t>0.

ob=0: boundary constant for an optional moving boundary. At time t, the boundary is located at ob*√t. Must be positive if eq is radial.

Examples

julia> D(θ) = θ^4
D (generic function with 1 method)

julia> prob = Fronts.CauchyProblem(D, b=2, d_dob=-0.1)
⎧ ∂θ/∂t = ∂(D(θ)*∂θ/∂r)/∂r, r>0,t>0
⎨ θ(0,t) = 2, t>0
⎩ √t*∂θ/∂r(0,t) = -0.1, t>0

See also: Equation

Fronts.FiniteProblem — Type

FiniteProblem{Eq<:Equation}

Abstract type for problems defined in finite domains.

Fronts.FiniteDirichletProblem — Type

FiniteDirichletProblem(eq, rstop[, tstop]; i, b) <: FiniteProblem

Models eq in the domain 0 ≤ r ≤ rstop with initial condition i and boundary condition b.

Arguments

eq: equation to solve.

rstop: length of the domain.

tstop=Inf: final time. If Inf, the solution is computed until the steady state is reached.

Keyword arguments

i: initial value.

b: boundary value.

Fronts.FiniteReservoirProblem — Type

FiniteReservoirProblem(eq, rstop, tstop; i, b, capacity) <: FiniteProblem

Models eq in the domain 0 ≤ r ≤ rstop with initial condition i and a reservoir boundary condition.

Arguments

eq: equation to solve.

rstop: length of the domain.

tstop=Inf: final time. If Inf, the solution is computed until the steady state is reached.

Keyword arguments

i: initial value.

b: boundary value.

capacity: reservoir capacity.

Fronts.monotonicity — Function

monotonicity(prob) -> Int

Whether the solution to prob must be decreasing (-1), constant (0) or increasing (+1) in r.